pyproject.toml
[tool.poetry]
name = "building-microservice-with-fastapi"
version = "0.1.0"
description = ""
authors = ["Bohdan Obruch <bodan1992@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
fastapi = "^0.111.1"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pydantic = "^2.8.2"
fastapi-pagination = "^0.12.26"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.31"
sqlmodel = "^0.0.21"
faker = "^26.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "smoke: mark a test as a smoke test."
]